requirements.txt
pytorch==0.3.0
torchvision==0.2.0
tqdm==4.23.0